AI Assistant Python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
bcrypt==4.1.2
passlib==1.7.4
pandas==2.1.3
openpyxl==3.1.2
itsdangerous==2.1.2